+++ source/js/modules/slider.js
@@ -1,5 +1,5 @@
 import Swiper from "swiper";
-import ThemeControl from "./themeControl";
+import ThemeControl from "./theme-control";
 const themeControl = new ThemeControl();
 
 export default () => {
