+++ source/scss/blocks/slider.scss
@@ -96,6 +96,26 @@
   max-width: 34rem;
   font-size: 1.8rem;
   line-height: 2rem;
+  opacity: 0;
+  transform: translateX(20px);
+
+  .swiper-slide-active & {
+    opacity: 1;
+    transform: translateX(0);
+    transition-property: transform, opacity;
+    transition-duration: 0.4s, 1s;
+    transition-timing-function: ease-out, ease;
+    transition-delay: 0.3s, 0.3s;
+  }
+
+  .swiper-slide-next & {
+    opacity: 1;
+    transform: translateX(0);
+    transition-property: transform, opacity;
+    transition-duration: 0.7s, 1s;
+    transition-timing-function: ease-out, ease;
+    transition-delay: 0.6s, 0.6s;
+  }
 
   @media (min-width: $stop-scaling) {
     margin-top: 36px;
